server/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
EbookLib==0.19
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eng_to_ipa==0.0.2
Faker==37.6.0
Flask==3.1.2
fonttools==4.58.5
ftfy==6.3.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
locate==1.1.1
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msgpack==1.1.0
murmurhash==1.0.13
numpy==2.2.6
packaging==25.0
pillow==11.3.0
preshed==3.0.10
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wordfreq==3.1.1
wrapt==1.17.2
